Revert "Test/error handling"
