+++ .gitignore
@@ -78,7 +78,6 @@ web_modules/
 
 # dotenv environment variable files
 .env
-.test.env
 .env.development.local
 .env.test.local
 .env.production.local
