--- src/errors/ExistentEmailError.ts
@@ -1,9 +1,0 @@
-export class ExistentEmailError extends Error {
-  public details?: string;
-  public readonly code: number;
-
-  constructor(message: string) {
-    super(message);
-    this.code = 409;
-  }
-}
