+++ test/index.spec.ts
@@ -7,16 +7,6 @@ import { initApolloServer } from '../src/apollo-server';
 import { AppDataSource } from '../src/data-source';
 import { User } from '../src/entity/User';
 import { comparePassword } from '../src/encryptPassword';
-import { mutationCreateUser, queryUser } from './queries';
-import { UserInput } from './resolvers';
-import { ExistentEmailError } from '../src/errors/ExistentEmailError';
-import { PasswordInvalidError } from '../src/errors/PasswordInvalidError';
-
-interface ApolloErrorFormat {
-  message: string;
-  code: number;
-  details: string;
-}
 
 const connection = axios.create({ baseURL: `http://localhost:${process.env.APOLLO_SERVER_PORT}` });
 
@@ -27,21 +17,44 @@ before(async () => {
 
 describe('Users Query Test', () => {
   it('Should bring number of users from database if server is online', async () => {
-    const result = (await queryUser(connection)).data.users as UserInput[];
-    expect(result.length).to.be.gt(0);
+    const query = `query User{
+      users{
+        id
+        name
+        email
+        birthdate
+      }
+    }`;
+
+    const result = await connection.post('/graphql', { query });
+    expect(result.data.data.users.length).to.be.eq(5);
   });
 });
 
 describe('CreateUser Mutation Test', () => {
-  it('Should newUser inserted in database and graphql result be equal to input', async () => {
+  it('Should result have the same data that was informed in the input', async () => {
+    const query = `mutation CreateUser($input: UserInput){
+      createUser(input: $input) {
+        id
+        name
+        email
+        birthdate
+      }
+    }`;
+
     const input = {
       name: 'MochaUser',
       email: 'mochauser@email.com',
       password: 'mochauserPassword1',
-      birthdate: '01-01-1993',
+      birthdate: '01-01-1900',
     };
 
-    const createdUser = (await mutationCreateUser(connection, input)).data.createUser as UserInput;
+    const result = await connection.post('/graphql', {
+      query,
+      variables: { input },
+    });
+
+    const createdUser = result.data.data.createUser;
     const userInDatabase = await AppDataSource.getRepository(User).findOneBy({ email: input.email });
 
     const isSamePassword = await comparePassword(input.password, userInDatabase.password);
@@ -62,42 +75,5 @@ describe('CreateUser Mutation Test', () => {
     after(async () => {
       await AppDataSource.getRepository(User).delete({ email: input.email });
     });
-  });
-
-  it('Should give an error when creating a user with an already existent email', async () => {
-    const input = {
-      name: 'MochaUser2',
-      email: 'mochauser2@email.com',
-      password: 'mochauserPassword2',
-      birthdate: '01-01-1995',
-    };
-
-    await mutationCreateUser(connection, input);
-    const apolloErrors = (await mutationCreateUser(connection, input)).errors as ApolloErrorFormat[];
-    const mailError = new ExistentEmailError('');
-    const hasDuplicatedEmailError = apolloErrors.some((error) => error.code === mailError.code);
-
-    expect(apolloErrors.length).to.be.gt(0);
-    expect(hasDuplicatedEmailError).to.be.true;
-
-    after(async () => {
-      await AppDataSource.getRepository(User).delete({ email: input.email });
-    });
-  });
-
-  it('Should give an error when creating a user with an invalid password', async () => {
-    const input = {
-      name: 'MochaUser3',
-      email: 'mochauser3@email.com',
-      password: 'mochauserPassword',
-      birthdate: '01-01-2000',
-    };
-
-    const apolloErrors = (await mutationCreateUser(connection, input)).errors as ApolloErrorFormat[];
-    const passwordError = new PasswordInvalidError('');
-    const hasInvalidPasswordError = apolloErrors.some((error) => error.code === passwordError.code);
-
-    expect(apolloErrors.length).to.be.gt(0);
-    expect(hasInvalidPasswordError).to.be.true;
   });
 });
