+++ tsconfig.json
@@ -4,12 +4,11 @@
          "es5",
          "es6"
       ],
-      "target": "ES2020",
+      "target": "es5",
       "module": "CommonJS",
       "moduleResolution": "node",
       "outDir": "./dist",
-      // "rootDir": "./src",
-      "rootDirs": ["./src","./test"],
+      "rootDir": "./src",
       "emitDecoratorMetadata": true,
       "experimentalDecorators": true,
       "strictPropertyInitialization": false,
